feat: remove auth
