+++ src/app/[locale]/(root)/history/loading.tsx
