+++ src/app/[locale]/(root)/variables/loading.tsx
