+++ src/app/[locale]/(root)/variables/page.tsx
