+++ src/app/[locale]/(utility)/layout.tsx
@@ -1,0 +1,7 @@
+export default function PublicLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return children;
+}
