+++ src/components/shared/Navigation.tsx
@@ -1,29 +1,12 @@
 'use client';
 
-import { LogOut } from 'lucide-react';
-import { useTranslations } from 'next-intl';
-import { useRouter, useSelectedLayoutSegment } from 'next/navigation';
-
-import { useAuth } from '@/hooks';
-import { supabase } from '@/lib';
 import { AppRoutes } from '@/services';
-import { Button, Link } from '@heroui/react';
+import { Link } from '@heroui/react';
 
 import { Logo } from '../ui';
-import { AuthButtons } from './AuthButtons';
 import { LocaleSwitcher } from './LocaleSwitcher';
 
 export function Navigation() {
-  const t = useTranslations('Navigation');
-  const segment = useSelectedLayoutSegment();
-  const user = useAuth();
-  const router = useRouter();
-
-  const handleLogout = () => {
-    supabase.auth.signOut();
-    router.replace(AppRoutes.HOME);
-  };
-
   return (
     <div className="sticky top-0 z-50">
       <div className="max-w-8xl mx-auto">
@@ -43,27 +26,6 @@ export function Navigation() {
             <div>
               <LocaleSwitcher />
             </div>
-            {segment === AppRoutes.SIGN_IN.slice(1) ||
-            segment === AppRoutes.SIGN_UP.slice(1) ||
-            user ? (
-              <>
-                <Link
-                  href={AppRoutes.HOME}
-                  className="px-4 py-2 text-slate-600 hover:bg-slate-100 rounded-lg transition-colors text-sm"
-                >
-                  {t('home')}
-                </Link>
-                <Button
-                  onPress={handleLogout}
-                  className="px-4 py-2 bg-red-50 text-red-600 hover:bg-red-100 rounded-lg transition-colors flex items-center gap-1.5"
-                >
-                  <LogOut className="w-5 h-5" />
-                  {t('Sign Out')}
-                </Button>
-              </>
-            ) : (
-              <AuthButtons />
-            )}
           </div>
         </div>
       </div>
